Move Regexp checking from ParserSupport to Lexer
While not really related to Lexing, this is a component that is shared between Ripper and the main parser, and that seemed like the lesser evil.
